pyproject.toml
[tool.poetry]
name = "srs-sqlite"
version = "0.1.0"
description = "A simple SRS app using Markdown/HandsOnTable/SQLite"
authors = ["Pacharapol Withayasakpunt <patarapolw@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/patarapolw/srs-sqlite"
homepage = "https://github.com/patarapolw/srs-sqlite"

keywords = ['srs', 'handsontable', 'flashcards', 'spaced-repetition']

[tool.poetry.dependencies]
python = ">=3"
flask = "^1.0"
flask_sqlalchemy = "^2.3"
python-dateutil = "^2.7"
click = "^6.7"

[tool.poetry.dev-dependencies]
jupyter = "^1.0"
notebook = "^5.6"
pyexcel = "^0.5.8"
pyexcel-xls = "^0.5.7"

[tool.poetry.scripts]
srs = 'srs_sqlite.__main__:cli'
